Made it work depending on username input.

+++ index.js
@@ -4,33 +4,41 @@ function deleteCar()
     deleteButton.type = "submit"
     deleteButton.value = "Delete Car"
     deleteButton.style.background = "#DB1200"
-    document.body.appendChild(deleteButton)
+    carDetails.appendChild(deleteButton)
 
 
 
     deleteButton.addEventListener("click", () => 
     {
-        fetch(`${url}/${currentCar.id}`, 
+        if(usernameInput.value.toLowerCase() === currentCar.user)
         {
-            method: "DELETE"
-        })
-        .then(resp => 
+            fetch(`${url}/${currentCar.id}`, 
             {
-                if(resp.ok)
-                {
-                    carsCopy = carsCopy.filter(car => 
-                        {
-                            return currentCar.id !== car.id
-                        })
-                    displayCarDetails(carsCopy[0])
-                    createNavBar(carsCopy)
-                    alert("Car deleted!")
-                }
-
-                else
-                {
-                    alert("Error: Unable to delete car!")
-                }
+                method: "DELETE"
             })
+            .then(resp => 
+                {
+                    if(resp.ok)
+                    {
+                        carsCopy = carsCopy.filter(car => 
+                            {
+                                return currentCar.id !== car.id
+                            })
+                        displayCarDetails(carsCopy[0])
+                        createNavBar(carsCopy)
+                        alert("Car deleted!")
+                    }
+                    else
+                    {
+                        alert("Error: Unable to delete car!")
+                    }
+            })
+        }
+        
+        else
+        {
+            alert("Error: Incorrect username!  Please try again!")
+        }
+    carForm.reset()
     })
 }
